Update tabs.js
When you have a paginator the current tab gets forgotten. This is due to the fact the current tab is stored in localstorage  by using the window.location.pathname. This changes when you use the pager. /page:2 gets added and so the code doesn't know what the state was. 
In this pr i stripped the page param from the url and its all working correctly.